requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiosqlite==0.19.0
aiohttp==3.9.5
anthropic==0.49.0